// apps/sales-copilot/src/generated/models/Crf5c_businessinsightsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Crf5c_businessinsightscrf5c_isactive = {
  0: 'No',
  1: 'Yes'
} as const;
export type Crf5c_businessinsightscrf5c_isactive = keyof typeof Crf5c_businessinsightscrf5c_isactive;
export const Crf5c_businessinsightscrf5c_referencetype = {
  995340000: 'client',
  995340001: 'opportunity'
} as const;
export type Crf5c_businessinsightscrf5c_referencetype = keyof typeof Crf5c_businessinsightscrf5c_referencetype;
export const Crf5c_businessinsightscrf5c_type = {
  995340000: 'warning',
  995340001: 'info',
  995340002: 'success'
} as const;
export type Crf5c_businessinsightscrf5c_type = keyof typeof Crf5c_businessinsightscrf5c_type;
export const Crf5c_businessinsightsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Crf5c_businessinsightsstatecode = keyof typeof Crf5c_businessinsightsstatecode;
export const Crf5c_businessinsightsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Crf5c_businessinsightsstatuscode = keyof typeof Crf5c_businessinsightsstatuscode;

export interface Crf5c_businessinsightsBase {
  crf5c_businessinsightid: string;
  crf5c_detailsjson: string;
  crf5c_displayorder: string;
  crf5c_generatedon: string;
  crf5c_isactive: Crf5c_businessinsightscrf5c_isactive;
  crf5c_ownerid: string;
  crf5c_rationale: string;
  crf5c_referenceidsjson: string;
  crf5c_referencetype: Crf5c_businessinsightscrf5c_referencetype;
  crf5c_summary: string;
  crf5c_title: string;
  crf5c_type: Crf5c_businessinsightscrf5c_type;
  crf5c_validuntil: string;
  importsequencenumber?: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Crf5c_businessinsightsstatecode;
  statuscode?: Crf5c_businessinsightsstatuscode;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
}

export interface Crf5c_businessinsights extends Crf5c_businessinsightsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  crf5c_isactivename?: string;
  crf5c_referencetypename?: string;
  crf5c_typename?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}